Move version generation code to Exanite.Arpg assembly
