+++ Assets/Exanite.Arpg/Versioning/Internal.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 82fc279086d1989458b6b579fc1e8604
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
