+++ Assets/Exanite.Arpg/Versioning/Internal/Git.cs.meta
